requirements.txt
aniso8601==9.0.1
arabic-reshaper==2.1.3
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-RESTful==0.3.9
future==0.18.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
